feat: 모듈 구조 개선
